removed ContirutorSearchContainer
