+++ src/domain/community/user/ContributorsPage.tsx
@@ -1,6 +1,5 @@
-import { useState, ChangeEvent } from 'react';
+import { useState, ChangeEvent, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import ContributorsSearchContainer from '../contributor/ContributorsSearch/ContributorsSearchContainer';
 import { InputAdornment, OutlinedInput } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
 import { debounce } from 'lodash';
@@ -14,6 +13,39 @@ import { GroupOutlined } from '@mui/icons-material';
 import TopLevelPageBreadcrumbs from '@/main/topLevelPages/topLevelPageBreadcrumbs/TopLevelPageBreadcrumbs';
 import GroupOutlinedIcon from '@mui/icons-material/GroupOutlined';
 import BreadcrumbsItem from '@/core/ui/navigation/BreadcrumbsItem';
+import {
+  useContributorsPageOrganizationsQuery,
+  useContributorsPageUsersQuery,
+  useContributorsVirtualInLibraryQuery,
+} from '@/core/apollo/generated/apollo-hooks';
+import {
+  ContributorsPageOrganizationsQuery,
+  ContributorsPageOrganizationsQueryVariables,
+  ContributorsPageUsersQuery,
+  ContributorsPageUsersQueryVariables,
+  OrganizationContributorFragment,
+  OrganizationVerificationEnum,
+  UserContributorFragment,
+} from '@/core/apollo/generated/graphql-schema';
+import { arrayShuffle } from '@/core/utils/array.shuffle';
+import usePaginatedQuery from '@/domain/shared/pagination/usePaginatedQuery';
+import { VirtualContributorModelBase } from '../virtualContributor/model/virtual.contributor.base.model';
+import { ApolloError } from '@apollo/client';
+
+export interface VirtualContributors {
+  items: VirtualContributorModelBase[] | undefined;
+  loading: boolean;
+}
+
+export interface PaginatedResult<T> {
+  items: T[] | undefined;
+  hasMore: boolean | undefined;
+  pageSize: number;
+  firstPageSize: number;
+  loading: boolean;
+  error?: ApolloError;
+  fetchMore: (itemsNumber?: number) => Promise<void>;
+}
 
 const ContributorsPage = () => {
   const { t } = useTranslation();
@@ -22,9 +54,13 @@ const ContributorsPage = () => {
   const [searchEnabled] = useState(false);
 
   const [searchTerms, setSearchTerms] = useState('');
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [searchTermsDebounced, setSearchTermsDebounced] = useState('');
 
   const { isAuthenticated } = useUserContext();
+
+  const pageSize = ITEMS_PER_PAGE;
 
   const onSearchHandlerDebounced = debounce((value: string) => setSearchTermsDebounced(value), 500);
 
@@ -34,6 +70,85 @@ const ContributorsPage = () => {
   };
 
   const ribbon = useInnovationHubOutsideRibbon({ label: 'innovationHub.outsideOfSpace.contributors' });
+
+  const usersQueryResult = usePaginatedQuery<ContributorsPageUsersQuery, ContributorsPageUsersQueryVariables>({
+    useQuery: useContributorsPageUsersQuery,
+    options: {
+      fetchPolicy: 'cache-first',
+      nextFetchPolicy: 'cache-first',
+      skip: !isAuthenticated,
+    },
+    variables: {
+      withTags: true,
+      filter: { firstName: searchTerms, lastName: searchTerms, email: searchTerms },
+    },
+    pageSize: pageSize,
+    getPageInfo: result => result.usersPaginated.pageInfo,
+  });
+
+  const randomizedUsers = useMemo(() => {
+    // if the length changed, shuffle only the new portion of the array
+    // to avoid re-rendering the entire list
+    const users = usersQueryResult.data?.usersPaginated.users;
+
+    if (!users) {
+      return [];
+    }
+
+    const randomizedNewUsers = arrayShuffle(users.slice(-pageSize));
+
+    return [...users.slice(0, users.length - pageSize), ...randomizedNewUsers];
+  }, [usersQueryResult.data?.usersPaginated.users?.length]);
+
+  const users: PaginatedResult<UserContributorFragment> = {
+    items: randomizedUsers,
+    loading: usersQueryResult.loading,
+    hasMore: usersQueryResult.hasMore,
+    pageSize: usersQueryResult.pageSize,
+    firstPageSize: usersQueryResult.firstPageSize,
+    error: usersQueryResult.error,
+    fetchMore: usersQueryResult.fetchMore,
+  };
+
+  const organizationsQueryResult = usePaginatedQuery<
+    ContributorsPageOrganizationsQuery,
+    ContributorsPageOrganizationsQueryVariables
+  >({
+    useQuery: useContributorsPageOrganizationsQuery,
+    options: {
+      fetchPolicy: 'cache-first',
+      nextFetchPolicy: 'cache-first',
+    },
+    variables: {
+      status: OrganizationVerificationEnum.VerifiedManualAttestation,
+      filter: {
+        contactEmail: searchTerms,
+        displayName: searchTerms,
+        domain: searchTerms,
+        nameID: searchTerms,
+        website: searchTerms,
+      },
+    },
+    pageSize: pageSize,
+    getPageInfo: result => result.organizationsPaginated.pageInfo,
+  });
+
+  const organizations: PaginatedResult<OrganizationContributorFragment> = {
+    items: organizationsQueryResult.data?.organizationsPaginated.organization,
+    loading: organizationsQueryResult.loading,
+    hasMore: organizationsQueryResult.hasMore,
+    pageSize: organizationsQueryResult.pageSize,
+    firstPageSize: organizationsQueryResult.firstPageSize,
+    error: organizationsQueryResult.error,
+    fetchMore: organizationsQueryResult.fetchMore,
+  };
+
+  const { data: vcData, loading: loadingVCs } = useContributorsVirtualInLibraryQuery();
+
+  const virtualContributors = {
+    items: vcData?.platform.library.virtualContributors ?? [],
+    loading: loadingVCs,
+  };
 
   return (
     <TopLevelPageLayout
@@ -65,18 +180,12 @@ const ContributorsPage = () => {
             />
           </PageContentBlockSeamless>
         )}
-        <ContributorsSearchContainer searchTerms={searchTermsDebounced} pageSize={ITEMS_PER_PAGE}>
-          {({ users, organizations, virtualContributors }) => {
-            return (
-              <ContributorsView
-                usersPaginated={users}
-                showUsers={isAuthenticated}
-                organizationsPaginated={organizations}
-                virtualContributors={virtualContributors}
-              />
-            );
-          }}
-        </ContributorsSearchContainer>
+        <ContributorsView
+          usersPaginated={users}
+          showUsers={isAuthenticated}
+          organizationsPaginated={organizations}
+          virtualContributors={virtualContributors}
+        />
       </PageContentColumn>
     </TopLevelPageLayout>
   );
